requirements.txt
python-dotenv==1.0.1
speedtest-cli==2.1.3
